==> 02_processing_sc_RNA_and_preparation_deconvolution.Rmd <==
---
title: "02_processing_sc_RNA_and_preparation_deconvolution"
author: "Yayuan Zhu"
date: "2024-08-19"
output: html_document
---gc()
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 2.1 Introduction

------------------------------------------------------------------------

**Goal**: Format bulk RNA-seq dataset according to the input specifications of Scaden.

**Overview workflow**:

-   Processing scRNA-seq dataset depending on which one to use as a reference during the deconvolution analysis.

    -   Athero-Express: remove cells with n \< 50 and group cell types into 7 major cell types.

    -   Tabula Sapiens: retain data from selected organs and group cell types into 10, 13, and 17 major cell types.

-   Match the genes of those in the bulk RNA-seq and scRNA-seq dataset.

-   Restructure the files into the desired format of Scaden.

**Note**: Users can adapt the code in chunks or lines marked with `# MODIFY HERE`. Other parts not annotated, such as file import and export, may also require changes to fit ones situation.

**Libraries**:

```{r}
library(tidyverse)
library(magrittr)
library(reshape2)

library(Seurat)

library(scCustomize)
```

# 2.2 Processing scRNA-seq reference

To use the AtheroExpress plaque scRNA-seq dataset as the reference for the cellular deconvolution, continue to Section 2.2.1. For the use of the Tabula Sapiens dataset as the reference, move to Section 2.2.2.

## 2.2.1 AtheroExpress plaque

```{r}
scAEplaque <- subset(read_rds("data/raw_data/full.43p_10X.integrated.cleaned.archetypes.seurat.RDS"), Tissue != "PBMC")

scAEplaque@meta.data %<>%
  mutate("cluster25" = as.character(Idents(scAEplaque)))

# Filter out low frequency cell types
ctFreq <- as_tibble(table(original = scAEplaque@meta.data$cluster25))

keepCellTypes <- factor(ctFreq$original[ctFreq$n >= 50])

ctFreq <- ctFreq[ctFreq$original %in% keepCellTypes,]
scAEplaqueFilterFreq <- subset(scAEplaque, idents = keepCellTypes)
```

```{r}
# Group into 7 major cell types
patternCellTypes <- list("Dendritic cell" = "(?i)cDC1|cDC2",
                         "Endothelial cell" = "(?i)endothelial",
                         "Macrophage" = "(?i)macrophage",
                         "Mast cell" = "(?i)mast",
                         "Memory B cell" = "(?i)\\sB\\scell",
                         "Smooth muscle cell" = "(?i)smooth",
                         "T/NK cell" = "(?i)(\\sT\\scell)|(NK\\scell)")

cluster <- lapply(patternCellTypes, function(pattern) {
  str_subset(ctFreq$original, pattern)
})

# Adapt cell type overview dataframe
ctFreq$cluster7 <- ctFreq$original

for(name in names(cluster)) {
  ctFreq$cluster7[ctFreq$original %in% cluster[[name]]] <- name
}

# Adapt metadata of the Seurat object
scAEplaqueFilterFreq@meta.data$cluster7 <- scAEplaqueFilterFreq@meta.data$cluster25

for(name in names(cluster)) {
  scAEplaqueFilterFreq@meta.data$cluster7[scAEplaqueFilterFreq@meta.data$cluster7 %in% cluster[[name]]] <- name
}
```

## 2.2.2 Tabula Sapiens

Due to memory limitations of the device used (with 16 GB of RAM and an AMD Ryzen 5 processor), certain steps are implemented during the processing of the Tabula Sapiens dataset. These steps include retaining organs with cardiovascular relevance or those expected to contribute significantly to total cfRNA, as well as limiting the dataset to 1000 cells of each cell type after the desired major cell type grouping has been applied.

```{r}
scTS <- read_rds("data/raw_data/Tabula_Sapiens_all_cells.rds")

Idents(scTS) <- scTS@meta.data$cell_ontology_class

# Only retain major contributing organs
organsToInclude <- c("Blood", "Liver", "Lymph_Node", "Fat", "Bone_Marrow", "Heart", "Muscle", "Pancreas", "Large_Intestine", "Lung", "Small_Intestine", "Vasculature", "Kidney", "Spleen")

scTSMetadata <- scTS@meta.data %>%
  filter(tissue_in_publication %in% organsToInclude &
           !(tissue_in_publication == "Heart" & cell_ontology_class == "hepatocyte") &
           !(cell_ontology_class == "endothelial cell" & ! (tissue_in_publication %in% c("Vasculature", "Fat"))))

scTSMetadata$cell_ontology_class <- as.character(scTSMetadata$cell_ontology_class)

gc()
```

```{r}
# Group cells into major cell types
ctFreq <- as_tibble(table(original = scTSMetadata$cell_ontology_class))

patternCellTypes <- list(
  "Monocyte" = "(?i)dendritic|monocyte|macrophage|langerhans",
  "Mesenchymal cell/fibroblast" = "(?i)fibroblast|stellate|adventitial|mesenchymal",
  "Skeletal muscle cell" = "(?i)[^hc] muscle",
  "B cell" = "(?i)(^|\\b)b\\scell|plasma[^c]",
  "Platelet" = "(?i)platelet",
  "Basophil/mast cell" = "(?i)basophil|mast",
  "Neutrophil" = "(?i)neutrophil",
  "Erythrocyte" = "(?i)erythrocyte",
  "Cardiac muscle cell" = "(?i)cardiac muscle",
  "Pericyte/smooth muscle cell" = "(?i)smooth|myometrial|pericyte",
  "T/NK cell" = "(?i)(nk)|(natural)|((\\A|\\W)t cell)",
  "Kidney epithelial cell" = "(?i)kidney",
  "Enterocyte" = "(?i)^(enterocyte|mature enterocyte)",
  "Hepatocyte" = "(?i)hepatocyte",
  "Pancreatic epithelial cell" = "(?i)pancreatic (ductal|acinar)",
  "Pneumocyte" = "(?i)pneumocyte",
  "Endothelial cell" = "(?i)^endothelial cell\\Z|endothelial cell of [^h]|lymphatic|vein endo|artery endo|cardiac endo|capillary endo")

cluster <- lapply(patternCellTypes, function(pattern) {
  str_subset(ctFreq$original, pattern)
})

# Modify cell type overview dataframe
ctFreq$cluster17 <- ctFreq$original

for(name in names(cluster)) {
  ctFreq$cluster17[ctFreq$original %in% cluster[[name]]] <- name
}

epithelialCells <- c("Kidney epithelial cell", "Enterocyte", "Hepatocyte", "Pancreatic epithelial cell", "Pneumocyte")

plaqueCells <- c("Monocyte", "Mesenchymal cell/fibroblast", "B cell", "Platelet", "Basophil/mast cell", "Neutrophil", "Erythrocyte", "Pericyte/smooth muscle cell", "T/NK cell", "Endothelial cell")

cells7 <- c("Endothelial cell", "Epithelial cell", "Monocyte", "T/NK cell", "Neutrophil", "Mesenchymal cell/fibroblast", "Pericyte/smooth muscle cell")

ctFreq %<>%
  mutate(cluster13 = ifelse(cluster17 %in% epithelialCells, "Epithelial cell", cluster17),
         cluster10 = ifelse(cluster13 %in% plaqueCells, cluster13, NA),
         cluster7 = ifelse(cluster13 %in% cells7, cluster13, NA))

# Modify metadata of the Seurat object
scTSMetadata$cluster17 <- scTSMetadata$cell_ontology_class

for(name in names(cluster)) {
  scTSMetadata$cluster17[scTSMetadata$cluster17 %in% cluster[[name]]] <- name
}

scTSMetadata$cluster17[!(scTSMetadata$cluster17 %in% names(cluster))] <- NA

scTSMetadata %<>%
  mutate(cluster13 = ifelse(cluster17 %in% epithelialCells, "Epithelial cell", cluster17),
         cluster10 = ifelse(cluster13 %in% plaqueCells, cluster13, NA),
         cluster7 = ifelse(cluster13 %in% cells7, cluster13, NA)) %>%
  mutate_at(c("cluster17", "cluster13", "cluster10", "cluster7"), factor)
```

```{r}
scTS@meta.data <- scTSMetadata
```

Below, a maximum of 1000 cells of each cell types from the chosen cluster will be included at random. Multiple Seurat objects can be created at ones of this reference in order to assess whether the randomization has an effect on the outcomes. Findings from the study indicate that this can be considered irrelevant.

```{r}
clusterToUse <- "cluster13" # MODIFY HERE
nSeurat <- seq_len(1) # MODIFY HERE; number of Seurat objects to create

ctFreqNew <- as_tibble(table(Celltype = scTSMetadata[[clusterToUse]]))

ctAbove1000 <- ctFreqNew$Celltype[ctFreqNew$n > 1000]
cellsBelow1000 <- scTSMetadata %>%
  subset(.[[clusterToUse]] %in% setdiff(ctFreqNew$Celltype, ctAbove1000)) %>%
  rownames()

for(n in nSeurat) {
  sampledCells <- c()
  
  for(cell in ctAbove1000) {
    rows <- scTSMetadata %>%
      subset(.[[clusterToUse]] == cell) %>%
      sample_n(1000) %>%
      rownames()
    
    sampledCells <- c(sampledCells, rows)
  }
  name <- paste0("cellsToInclude", n)
  
  assign(name, c(cellsBelow1000, sampledCells))
}

for(n in nSeurat) {
  list <- get(paste0("cellsToInclude", n))
  objName <- paste0("scTSCluster", gsub("cluster", "", clusterToUse), n)
  
  obj <- subset(scTS, cells = list)
  Idents(obj) <- obj@meta.data[[clusterToUse]]

  assign(objName, obj)
}
```

```{r}
rm(obj, scTS)
gc()
```

# 2.3 Integration of bulk RNA-seq and scRNA-seq data

------------------------------------------------------------------------

```{r}
bulkToMatch <- read_rds("data/preprocessed_data/bAEccfRNACombine.rds") # MODIFY HERE; RDS file from previous R Markdown

aeRefToUse <- FALSE # MODIFY HERE; using Athero-Express as reference for deconvolution = TRUE

# Filter bulk RNA-seq data on overlapping genes
if(aeRefToUse) {
  genesToKeep <- intersect(rownames(scAEplaqueFilterFreq), bulkToMatch$symbol)
  
  bulkFinal <- bulkToMatch %>%
    filter(symbol %in% genesToKeep) %>%
    column_to_rownames("symbol") %>%
    select(-ensembl_gene_id)
} else {
  genesToKeep <- intersect(rownames(get(paste0("scTS", str_to_title(clusterToUse), 1))), bulkToMatch$ensembl_gene_id)
  
  geneMap <- bulkToMatch %>%
    filter(ensembl_gene_id %in% genesToKeep) %>%
    select(ensembl_gene_id, symbol)
  
  map <- setNames(geneMap$symbol, geneMap$ensembl_gene_id)
  
  bulkFinal <- bulkToMatch %>%
    filter(ensembl_gene_id %in% genesToKeep) %>%
    column_to_rownames("symbol") %>%
    select(-ensembl_gene_id)
}

# Filter scRNA-seq data on overlapping genes
if(aeRefToUse) {
  scAEplaqueFilterGenes <- subset(scAEplaqueFilterFreq, features = genesToKeep)
} else {
  for(n in nSeurat) {
    obj <- get(paste0("scTSCluster", gsub("cluster", "", clusterToUse), n))
    objSubset <- subset(obj, features = genesToKeep)
  
    newObjName <- paste0("scTSClusterFilterGenes", gsub("cluster", "", clusterToUse), n)
    assign(newObjName, objSubset)
  }
}

gc()
```

# 2.4 Format files for cellular deconvolution

```{r}
# Generate count files with gene symbols
if(aeRefToUse) {
  scAECounts <- scAEplaqueFilterGenes[["RNA"]]$counts %>%
    as.data.frame() %>%
    t()
} else {
  for(n in nSeurat) {
    obj <- get(paste0("scTSClusterFilterGenes", gsub("cluster", "", clusterToUse), n))
    countsName <- paste0("scTSCounts", gsub("cluster", "", clusterToUse), n)
    
    counts <- obj[["RNA"]]$counts %>%
    as.data.frame() %>%
    t()
    
    colnames(counts) <- map[colnames(counts)]
    assign(countsName, counts)
    }
}

# Generate cell type files
if(aeRefToUse) {
  scAECellTypes <- scAEplaqueFilterGenes@meta.data %>%
    select(cluster7) %>%
    rename("Celltype" = "cluster7")
} else {
  for(n in nSeurat) {
    obj <- get(paste0("scTSClusterFilterGenes", gsub("cluster", "", clusterToUse), n))
    celltypesName <- paste0("scTSCellTypes", gsub("cluster", "", clusterToUse), n)
    
    celltypes <- Idents(obj) %>%
      as.data.frame() %>%
      rename(., Celltype = `.`)
    
    assign(celltypesName, celltypes)
    }
}

table(rownames(scTSCounts171) == rownames(scTSCellTypes171)) # MODIFY HERE
```

```{r}
patternBulk <- "rTS_bAEccfRNACombine_c17_v4" # MODIFY HERE
patternRef <- "rTS_bAEccfRNACombine_c17_v4" # MODIFY HERE

if(aeRefToUse) {
  cellTypesToExport <- "scAECellTypes"
  countsToExport <- "scAECounts"
} else {
  cellTypesToExport <- "scTSCellTypes171" # MODIFY HERE
  countsToExport <- "scTSCounts171" # MODIFY HERE
}

pathRef <- paste0("data/processed_data/scaden_sc_rna_ref_files/", patternRef, "/") # MODIFY HERE

if (!file.exists(pathRef)) {
  dir.create(pathRef)
  print("Folder created.")
} else {
  print("Folder already exists.")
}

write.table(bulkFinal,
            file = paste0("data/processed_data/scaden_bulk_rna_files/", patternBulk, "_bulk_data.txt"),
            sep = "\t")

write.table(get(cellTypesToExport),
            file = paste0(pathRef, patternRef, "_celltypes.txt"),
            sep = "\t", row.names = FALSE)

write.table(get(countsToExport),
            file = paste0(pathRef, patternRef, "_counts.txt"),
            sep = "\t", row.names = FALSE)
```
