==> 03_scaden_cellular_deconvolution.Rmd <==
---
title: "03_scaden_cellular_deconvolution"
author: "Yayuan Zhu"
date: "2024-01-31"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 3.1 Introduction

------------------------------------------------------------------------

**Goal**: Cellular deconvolution with Scaden in bash chunks of the desired datasets.

**Note**: Adapt the variables to import the user data.

# 3.2 Cellular deconvolution with Scaden

```{bash}
source functions/functions_bash.sh
set -e

patternRef="rTS_bAEccfRNACombine_c17_v3"
patternBulk="rTS_bAEccfRNACombine_c17_v3"

dirRef="data/processed_data/scaden_sc_rna_ref_files/${patternRef}"
dirBulk="data/processed_data/scaden_bulk_rna_files/${patternBulk}"
dirSimul="temp/scaden_simulated_data"
dirModel="temp/scaden_trained_models"
dirPrediction="output/scaden_predictions/${patternRef}"

createDir "${dirPrediction}"

# Setting both start and end to the same number will only run the loop once
start=1
end=10

for i in $(seq $start $end)
do
  scaden simulate -n 1000 --pattern "*_counts.txt" --data $dirRef --out $dirSimul

  scaden process "${dirSimul}/${patternRef}.h5ad" "${dirBulk}_bulk_data.txt" --processed_path "${dirSimul}/processed.h5ad"

  scaden train "${dirSimul}/processed.h5ad" --steps 5000 --model_dir $dirModel

  scaden predict --model_dir $dirModel "${dirBulk}_bulk_data.txt" --outname "${patternRef}_n${i}"

  mv "${patternRef}_n${i}" "${dirPrediction}"

  echo "Moved ${patternRef}_n${i} to ${dirPrediction}"

  emptyDir $dirSimul $i

  emptyDir $dirModel $i
done
```
