==> 04_analysis_cellular_deconvolution.Rmd <==
---
title: "04_analysis_cellular_deconvolution"
author: "Yayuan Zhu"
date: "2024-08-28"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 4.1 Introduction

------------------------------------------------------------------------

**Goal**: Analyze and visualize the predictions from cellular deconvolution analyses.

**Overview workflow**:

-   Obtain the average of 10 runs originating from the same dataset pair.

-   Calculate correlations with other predictions or normalized gene counts from the corresponding bulk RNA-seq dataset.

-   Visualize average predicted proportions and calculated correlations.

**Note**: Users can adapt the code in chunks or lines marked with `# MODIFY HERE`. Other parts not annotated, such as file import and export, may also require changes to fit ones situation.

**Libraries**:

```{r}
library(tidyverse)
library(edgeR)

library(ComplexHeatmap)

source("functions/functions_04_R.R")
source("functions/functions_global_R.R")
```

# 4.2 (Average) predicted cellular proportions

------------------------------------------------------------------------

```{r}
pathScaden <- "output/scaden_predictions/" # MODIFY HERE
fileBasePattern <- "rTS_bAEccfRNACombine_c17_v" # MODIFY HERE; base names of runs to import
nVersions <- 1 # MODIFY HERE; number or vector of the version(s)
nRuns <- seq_len(10) # MODIFY HERE; set of runs to import
```

```{r}
listPreds <- getScadenPred(pathScaden, fileBasePattern, nVersions, nRuns)

if(length(nVersions) == 1) {
  longPred <- listPreds[["avgLongPred"]] %||% listPreds[["longPred"]]
  widePred <- listPreds[["avgWidePred"]] %||% listPreds[["widePred"]]
} else {
  for(i in nVersions) {
    version <- paste0("version", i)
    
    longPred <- listPreds[[version]][["avgLongPred"]] %||% listPreds[[version]][["longPred"]]
    assign(paste0("longPred", i), longPred)
    
    widePred <- listPreds[[version]][["avgWidePred"]] %||% listPreds[[version]][["widePred"]]
    assign(paste0("widePred", i), widePred)
    
    rm(longPred, widePred)
  }
}
```

```{r}
saveRDS(longPred, file = paste0("output/export_files/rds_files/scaden_long_avg_prop_", fileBasePattern, nVersions, ".rds"))

saveRDS(widePred, file = paste0("output/export_files/rds_files/scaden_wide_avg_prop_", fileBasePattern, nVersions, ".rds"))
```

# 4.3 Calculating correlations

------------------------------------------------------------------------

Optionally, the imported data can be correlated with either another type of cellular deconvolution results, e.g. using another scRNA-seq reference or bulk RNA-seq input, in Section 4.3.1. For correlating the imported cellular deconvolution results with cell type gene markers, continue to Section 4.3.2.

## 4.3.1 Correlation between different predicted cellular proportions

```{r}
predToCorW <- readRDS(file = "output/export_files/rds_files/scaden_wide_avg_prop_rTS_bAEccfRNACombine_c17_v1.rds") # MODIFY HERE

predToCorL <- readRDS(file = "output/export_files/rds_files/scaden_long_avg_prop_rTS_bAEccfRNACombine_c17_v1.rds") # MODIFY HERE

if(fromScaden) {
  toCorrelateCol <- "avgProportion"
} else {
  toCorrelateCol <- "proportion"
}

matchPatients <- any(!predToCorW$patient %in% widePred$patient)

if(matchPatients) {
  cat("Matching patients...")
  matching <- intersect(predToCorW$patient, widePred$patient)
  
  filterPredToCorW <- predToCorW %>%
    filter(patient %in% matching)
  filterPredToCorL <- predToCorL %>%
    filter(patient %in% matching)
  
  filterWidePred <- widePred %>%
    filter(patient %in% matching)
  filterLongPred <- longPred %>%
    filter(patient %in% matching)
  
  filteredPatients <- TRUE
} else {
  cat("Patients are already matched.")
  filteredPatients <- FALSE
}
```

```{r}
# Generate dataframe for scatter plots
if(filteredPatients) {
  corPredScatter <- filterPredToCorL %>%
    left_join(filterWidePred, by = "patient")
} else {
  corPredScatter <- predToCorL %>%
    left_join(widePred, by = "patient")
}

corPredScatter %<>%
  rename(proportionToCor = !!sym(toCorrelateCol))
```

```{r}
# Generate datafrome for heatmap
if(filteredPatients) {
  print(table(filterPredToCorW$patient == filterWidePred$patient))
  
  predToCorWRow <- filterPredToCorW %>%
    arrange(patient) %>%
    column_to_rownames("patient")

  widePredRow <- filterWidePred %>%
    arrange(patient) %>%
    column_to_rownames("patient")
} else {
  print(table(predToCorW$patient == widePred$patient))
  
  predToCorWRow <- predToCorW %>%
    arrange(patient) %>%
    column_to_rownames("patient")

  
  widePredRow <- widePred %>%
    arrange(patient) %>%
    column_to_rownames("patient")
}

table(rownames(widePredRow) == rownames(predToCorWRow))

corPredHeatmap <- cor(widePredRow, predToCorWRow) # rows = widePredHeatmap; columns = predToCorrelateWideHeatmap
```

## 4.3.2 Correlation of predicted cellular proportions with cell type gene marker

```{r}
bulkGenesToCor <- read.table(file = "data/processed_data/scaden_bulk_rna_files/rTS_bAEccfRNACombine_bulk_data.txt", check.names = FALSE) # MODIFY HERE; bulk data used to obtain the predictions from Section 4.2

nCellTypes <- ncol(widePred) - 1

if(nCellTypes == 7) {
  geneMarkers <- list("Mast cell" = c("KIT", "TPSAB1", "CMA1"),
                      "Memory B cell" = c("MS4A1", "CD19"),
                      "T/NK cell" = c("CD3E", "IL7R", "KLRD1", "GNLY"),
                      "Macrophage" = c("CD68", "CD163"),
                      "Dendritic cell" = c("CD1A", "CD80"),
                      "Endothelial cell" = c("VWF", "ENG", "SELE", "CDH5"),
                      "Smooth muscle cell" = c("ACTA2", "MYH11"))
} else if(nCellTypes >= 10) {
  geneMarkers <- list("B cell" = c("MS4A1", "CD19"),
                      "Basophil/mast cell" = c("KIT", "TPSAB1", "CMA1"),
                      "Endothelial cell" = c("VWF", "ENG", "SELE", "CDH5"),
                      "Erythrocyte" = c("HBB", "GYPC"),
                      "Mesenchymal cell/fibroblast" = c("LUM", "PDGFRA"),
                      "Monocyte" = c("LYZ", "CD14", "CD16", "CD68", "CCR2", "CD86"),
                      "Neutrophil" = c("S100A8", "IL1R2", "CSF3R", "MPO"),
                      "Pericyte/smooth muscle cell" = c("ACTA2", "MYH11"),
                      "Platelet" = "PF4",
                      "T/NK cell" = c("CD3E", "IL7R", "KLRD1", "GNLY"))
}

if(nCellTypes >= 13) {
  addGeneMarkers <- list("Cardiac muscle cell" = c("TNNT2", "CNN1"),
                         "Epithelial cell" = c("EPCAM", "CDH1", "KRT8"),
                         "Skeletal muscle cell" = c("MYL1", "PAX7", "DES"))

  geneMarkers <- c(geneMarkers, addGeneMarkers)
}

if(nCellTypes == 17) {
  addGeneMarkers2 <- list("Enterocyte" = "FABP2",
                          "Hepatocyte" = "ALB",
                          "Kidney epithelial cell" = c("SLC5A12", "SLC34A1"),
                          "Pancreatic epithelial cell" = c("CTRC", "PNLIP"),
                          "Pneumocyte" = c("SFTPC", "AGER"))

  geneMarkers <- c(geneMarkers, addGeneMarkers2)
}

missingGenes <- unlist(geneMarkers)[!unlist(geneMarkers) %in% rownames(bulkGenesToCor)]
```

```{r}
bulkGeneMarkers <- bulkGenesToCor %>%
  filter(rownames(.) %in% unlist(geneMarkers))

widePredGeneMarkers <- column_to_rownames(widePred, "patient")

# Normalize bulk data before correlating with gene markers to account for sequencing depth and technical variation
# NOTE: bulk RNA-seq data of AEplaqueCombine and AEccfRNA(1/2/Combine) have already been normalized before cellular deconvolution
normBefore <- str_detect(fileBasePattern, "AEplaqueCombine|AEccfRNA|ACSccfRNA")

if(!normBefore) {
  cat("Normalizing bulk data...\n")
  dge <- DGEList(counts = bulkGeneMarkers)
  dgeNorm <- calcNormFactors(dge, method = "TMM")
  bulkGeneMarkersNorm <- cpm(dgeNorm) %>%
    t() %>%
    as.data.frame()
  
  print(table(rownames(widePredGeneMarkers) == rownames(bulkGeneMarkersNorm)))
  
  corGeneMarkers <- cor(widePredGeneMarkers, bulkGeneMarkersNorm)
} else {
  cat("Bulk has been normalized prior.\n")
  bulkGeneMarkersT <- as.data.frame(t(bulkGeneMarkers))
  
  print(table(rownames(widePredGeneMarkers) == rownames(bulkGeneMarkersT)))
  
  corGeneMarkers <- cor(widePredGeneMarkers, bulkGeneMarkersT)
}
```

# 4.4 Graphs

------------------------------------------------------------------------

```{r}
figPattern <- paste0(fileBasePattern, nVersions) # MODIFY HERE
```

## 4.4.1 Boxplots of predicted cellular proportion(s)

### 4.4.1.1 Boxplot of each run

```{r}
pdf(paste0("output/plots/boxplot_cell_type_proportion/pdf/scaden_", figPattern, "_multiple_runs_predProp.pdf"))

for(n in nRuns) {
  pRunsPred <- ggplot(listPreds[["listLongPred"]][[paste0("longRun", n)]], aes(x = reorder(cellType, proportion), y = proportion)) +
  geom_boxplot(fill = "#6CACD5", linewidth = 0.3, outlier.size = 1, outlier.shape = 1) +
  labs(title = figPattern,
       subtitle = paste0("Run ", n),
       x = "",
       y = "Estimated cell type proportion") +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, by = 0.1)) +
  theme_classic() +
  boxplotTheme()
  
  print(pRunsPred)
}

dev.off()
```

### 4.4.1.2 One (averaged) version

```{r}
pOnePred <- ggplot(longPred, aes(x = reorder(cellType, !!sym(colProp)), y = !!sym(colProp))) +
  geom_boxplot(fill = "#6CACD5", linewidth = 0.3, outlier.size = 1, outlier.shape = 1) +
  labs(title = figPattern,
       x = "",
       y = "Estimated cell type proportion") +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, by = 0.1)) +
  theme_classic() +
  boxplotTheme()

print(pOnePred)

ggsave(paste0("output/plots/boxplot_cell_type_proportion/tiff/scaden_", figPattern, "_predProp.tiff"),
     width = 9, height = 12, units = "cm", dpi = 300)
```

### 4.4.1.3 Groups

```{r}
# ---- Different versions ----
nVersionsCat <- paste(nVersions, collapse = "")

combinedLongPredVersions <- data.frame()

for(n in nVersions) {
  longPredN <- get(paste0("longPred", n)) %>%
    mutate(version = as.factor(paste0("v", n)))
  
  combinedLongPredVersions <- rbind(combinedLongPredVersions, longPredN)
}

rm(longPredN)

pVersions <- ggplot(combinedLongPredVersions, aes(x = reorder(cellType, !!sym(colProp)), y = !!sym(colProp))) +
  geom_boxplot(aes(fill = version), position = position_dodge(width = 0.8), width = 0.5, linewidth = 0.3, outlier.size = 1, outlier.shape = 1) +
  labs(title = paste0(fileBasePattern, nVersionsCat),
       x = "",
       y = "Estimated cell type proportion") +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, by = 0.1)) +
  scale_fill_manual(values = blueColorPal[["3"]]) + # MODIFY HERE; custom color palette options in functions_figures_R.R
  theme_classic() +
  boxplotTheme()

print(pVersions)

ggsave(paste0("output/plots/boxplot_cell_type_proportion/tiff/scaden_", fileBasePattern, nVersionsCat, "_predProp.tiff", collapse = ""), 
     width = 11, height = 12, units = "cm", dpi = 300)
```

```{r}
# ---- Different batches ----
batches <- readRDS("data/metadata/AEccfRNA_batch_patientID.rds") # MODIFY HERE

batchesLongPred <- longPred %>%
  left_join(batches, by = "patient")

pBatch <- ggplot(batchesLongPred, aes(x = reorder(cellType, !!sym(colProp)), y = !!sym(colProp))) +
  geom_boxplot(aes(fill = batch), position = position_dodge(width = 0.8), width = 0.5, linewidth = 0.3, outlier.size = 1, outlier.shape = 1) +
  labs(title = fileBasePattern,
       x = "",
       y = "Estimated cell type proportion") +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, by = 0.1)) +
  scale_fill_manual(values = blueColorPal[["2"]]) + # MODIFY HERE; custom color palette options in functions_figures_R.R
  theme_classic() +
  boxplotTheme()

print(pBatch)

ggsave("test.tiff", 
     width = 7, height = 12, units = "cm", dpi = 300)
```

```{r}
# ---- Different blood fractions ----
fractions <- read.delim("data/metadata/cfRNA_blood_pilot_annotation.txt", header = TRUE) %>%
  rename("patient" = "ID")

fractionsLongPred <- longPred %>%
  left_join(fractions, by = "patient")

pFractions <- ggplot(fractionsLongPred, aes(x = reorder(cellType, !!sym(colProp)), y = !!sym(colProp))) +
  geom_boxplot(aes(fill = tissue), position = position_dodge(width = 0.8), width = 0.5, linewidth = 0.3, outlier.size = 1, outlier.shape = 1) +
  labs(x = "",
       y = "Estimated cell type proportion") +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, by = 0.1)) +
  scale_fill_manual(values = blueColorPal[["3"]]) + # MODIFY HERE; custom color palette options in functions_figures_R.R
  theme_classic() +
  boxplotTheme()

print(pFractions)

ggsave("test4.tiff", 
     width = 15, height = 10, units = "cm", dpi = 300)
```

## 4.4.2 Correlation plots

### 4.4.2.1 Scatter plots

```{r}
path <- paste0("output/plots/correlation_scatter_plot/scaden_", figPattern, "_v2")

createDir(path)

for(i in (4:ncol(corPredScatter))) {
  tiff(paste0(path, "/scaden_", figPattern, "_v2_", i, "_scatter_cor.tiff"), width = 8, height = 8, units = "in", res = 300)
  
  colName <- colnames(corPredScatter)[i]
  
  pScatter <- ggplot(corPredScatter, aes(x = .data[[colName]], y = proportionToCor)) +
    geom_point(shape = 1, size = 1, stroke = 0.5) +
    geom_smooth(method = lm, linewidth = 0.5, color = "black", fill = "grey30") +
    facet_wrap(~ cellType, nrow = 5, scales = "free") +
    labs(x = paste0("AEplaque1_c7_v1", colName),
         y = "AEplaque1_c7_v2") + # MODIFY HERE; x corresponds to imported data in Section 4.3; y corresponds to imported data in Section 4.4.1
    theme_classic() +
    theme(strip.background = element_blank(),
          panel.grid = element_blank(),
          axis.title = element_text(size = 10),
          axis.title.x = element_text(margin = margin(t = 12)),
          axis.title.y = element_text(margin = margin(r = 12)),
          axis.text = element_text(size = 8),
          axis.text.x = element_text(margin = margin(t = 5)),
          axis.text.y = element_text(margin = margin(r = 5)))
  
  print(pScatter)
  
  dev.off()
}
```

### 4.4.2.2 Heatmap: between different predicted cellular proportions

```{r}
# NOTE: rows = widePredHeatmap; columns = predToCorrelateWideHeatmap
heatmapPred <- Heatmap(corPredHeatmap, col = heatmapPal, rect_gp = gpar(col = "white", lwd = 1),
        row_title = "rTS_bAEccfRNACominbine_com4_v1", column_title = "rTS_bAEccfRNACombine_c17_v1", # MODIFY HERE
        column_names_rot = 45, row_names_side = "left", row_dend_side = "right", column_title_side = "bottom",
        row_title_gp = gpar(fontsize = 8), column_title_gp = gpar(fontsize = 8), column_names_gp = gpar(fontsize = 6), row_names_gp = gpar(fontsize = 6),
        heatmap_width = unit(11, "cm"), heatmap_height = unit(11, "cm"),
        # Modify figure legend
        heatmap_legend_param = list(title = "Correlation",
                                    title_gp = gpar(fontsize = 8, fontface = "bold"), labels_gp = gpar(fontsize = 6)),
        # Add correlations as text
        layer_fun = function(j, i, x, y, width, height, fill) {
          grid.text(sprintf("%.1f", pindex(corPredHeatmap, i, j)), x, y, gp = gpar(fontsize = 6))
        }
)

print(heatmapPred)

tiff("output/plots/correlation_heatmap_predictions/scaden_rTS_bAEccfRNACombine_com4_c17.tiff", width = 13, height = 11.5, units = "cm", res = 300)

print(heatmapPred)

dev.off()
```

### 4.4.2.3 Heatmap: between predicted cellular proportions and cell type gene marker

```{r}
# NOTE: x-axis/columns = gene markers; y-axis/rows = cell types
epithelialCells <- c("Enterocyte", "Hepatocyte", "Kidney epithelial cell", "Pancreatic epithelial cell", "Pneumocyte")

heatmapGeneMarkers <- Heatmap(corGeneMarkers, col = heatmapPal, rect_gp = gpar(col = "white", lwd = 1),column_title = figPattern,
        column_names_rot = 45, row_names_side = "left", row_dend_side = "right",
        row_title_gp = gpar(fontsize = 8), column_title_gp = gpar(fontsize = 8, fontface = "bold"), column_names_gp = gpar(fontsize = 6), row_names_gp = gpar(fontsize = 6),
        # heatmap_width = unit(17, "cm"), heatmap_height = unit(5, "cm"), # For 10 cell types
        heatmap_width = unit(17, "cm"), heatmap_height = unit(8, "cm"),
        # Modify figure legend
        heatmap_legend_param = list(title = "Correlation",
                                    title_gp = gpar(fontsize = 8, fontface = "bold"), labels_gp = gpar(fontsize = 6)),
        # Add annotations of gene markers corresponding to cell types
        cell_fun = function(j, i, x, y, width, height, fill) {
          cellType <- rownames(corGeneMarkers)[i]
          geneMarker <- colnames(corGeneMarkers)[j]

          if(geneMarker %in% geneMarkers[[cellType]] ||
             (cellType %in% epithelialCells && geneMarker %in% geneMarkers[["Epithelial cell"]])) {
            grid.points(x, y, pch = 4, size = unit(0.3, "char"), gp = gpar(col = "black"))
          }
        }
)

print(heatmapGeneMarkers)

tiff(paste0("output/plots/correlation_heatmap_gene_markers/", figPattern, "_genes.tiff"),
     width = 19, height = 8, units = "cm", res = 300) # MODIFY HERE
print(heatmapGeneMarkers)

dev.off()
```
