==> 05_clinical_correlations.Rmd <==
---
title: "05_clinical_correlations"
author: "Yayuan Zhu"
date: "2024-08-28"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 5.1 Introduction

------------------------------------------------------------------------

**Goal**: Clinical analysis of cellular deconvolution outcomes with characteristics of Athero-Express cohort

**Overview workflow**:

-   Generate combined dataframe of predicted cell type proportions and clinical characteristics.

-   Perform simple statistical tests, GLM analysis, and Cox regression.

-   Visualize outcomes of clinical analysis.

**Note**: Users can adapt the code in chunks or lines marked with `# MODIFY HERE`. Other parts not annotated, such as file import and export, may also require changes to fit ones situation.

**Libraries**:

```{r}
library(tidyverse)
library(broom)
library(gtools) # Adds significance stars to tidy()

library(car) # vif
library(rstatix)
library(survival) # coxzph
library(ranger)
library(performance) # Hosmer-Lemeshow goodness-of-fit test
library(mice) # Impute missing data
library(haven)

library(patchwork)
library(Polychrome)

source("functions/functions_05_R.R")
source("functions/functions_global_R.R")
```

# 5.2 (Average) predicted cellular proportions and corresponding patient data

## 5.2.1 Athero-Express patient data

```{r}
patientData <- haven::read_sav("data/metadata/AE_patient_data_20230803.sav") %>%
  rename(patient = STUDY_NUMBER) # MODIFY HERE
patientData$patient <- gsub("^", "ae", patientData$patient)

# Add Symptoms.2g column
patientData$Symptoms.2g <- case_when(patientData$Symptoms.4g == 0 | patientData$Symptoms.4g == 1 ~ 0,
                                     patientData$Symptoms.4g == 2 | patientData$Symptoms.4g == 3 ~ 1)

patientData$Symptoms.2g <- factor(patientData$Symptoms.2g, levels = c(0, 1), labels = c("mild", "severe"))

# Add thrombus.bin column
patientData$thrombus.bin <- case_when(patientData$thrombus == 0 | patientData$thrombus == 1 ~ 0,
                                      patientData$thrombus == 2 | patientData$thrombus == 3 ~ 1,
                                      patientData$thrombus == -888 ~ NA)

patientData$thrombus.bin <- factor(patientData$thrombus.bin, levels = c(0, 1), labels = c("no/minor", "moderate/heavy"))

# Add hsCRP.bin column
patientData$hsCRP.bin <- case_when(patientData$hsCRP_plasma < 2 ~ 0,
                                   patientData$hsCRP_plasma >= 2 ~ 1,
                                   patientData$hsCRP_plasma >= 40 ~ NA)

patientData$hsCRP.bin <- factor(patientData$hsCRP.bin, levels = c(0, 1), labels = c("low", "high"))

# Add collagen.bin column
patientData$collagen.bin <- case_when(patientData$collagen == 0 | patientData$collagen == 1 ~ 0,
                                      patientData$collagen == 2 | patientData$collagen == 3 ~ 1,
                                      patientData$collagen == -888 | patientData$collagen == -999 ~ NA)

patientData$collagen.bin <- factor(patientData$collagen.bin, levels = c(0, 1), labels = c("no/minor", "moderate/heavy"))

#Add plaquephenotype.bin column
patientData$plaquephenotype.bin <- case_when(patientData$plaquephenotype == 1 ~ 0,
                                             patientData$plaquephenotype == 2 ~ NA,
                                             patientData$plaquephenotype == 3 ~ 1)

patientData$plaquephenotype.bin <- factor(patientData$plaquephenotype.bin, levels = c(0, 1), labels = c("fibrous", "atheromatous"))

# Columns to turn into factors
toFactor <- c("Gender", "DM.composite", "Hypertension.composite", "SmokerCurrent", "PAOD", "CAD_history", "Med.statin.derived", "Calc.bin", "Stroke_history", "Macrophages.bin", "SMC.bin", "IPH.bin")

patientData <- patientData %>%
  mutate_at(toFactor, as.factor)

# Turn EP_major into numeric
patientData$EP_major <- as.numeric(patientData$EP_major)
```

## 5.2.2 (Average) predicted cellular proportions

```{r}
proportionsAsFractions <- TRUE # MODIFY HERE; proportions as fractions are the default
zTransform <- TRUE # MODIFY HERE; set to FALSE for Section 5.4.1

pattern <- "rTS_bGTExHeart_c17_v1" # MODIFY HERE; base name of the predictions

pred <- readRDS(paste0("output/export_files/rds_files/scaden_wide_avg_prop_", pattern, ".rds")) %>% # MODIFY HERE; as wide format
  rename_with(.fn = ~ str_replace_all(., c(" " = "_", "/" = "_")),
              .cols = -patient)

# Optional modification
predModWide <- pred

  # < Convert proportions to percentages >
if(!proportionsAsFractions) {
  predModWide <- predModWide %>%
    mutate(across(-patient, ~ . * 100))
}

  # < Transform predicted cellular proportions to Z scores >
if(zTransform) {
  predModWide <- predModWide %>%
    mutate(across(-patient, ~ c(scale(.))))
}

predModLong <- predModWide %>%
  pivot_longer(cols = -patient, names_to = "cellType", values_to = "proportion")

# Add patient characteristics
predSympWide <- predModWide %>%
  left_join(patientData, by = "patient")

predSympLong <- predModLong %>%
  left_join(patientData, by = "patient")

# Add additional columns for bAEccfRNACombine
if(str_detect(pattern, "AEccfRNACombine")) {
  cat("Bulk RNA-seq data of AEccfRNACombine detected...")
  
  # Add batch ID
  batchesCcfRNA <- readRDS("data/metadata/AEccfRNA_batch_patientID.rds")
  
  predSympWide <- predSympWide %>%
    left_join(batchesCcfRNA, by = "patient")
  
  predSympLong <- predSympLong %>%
    left_join(batchesCcfRNA, by = "patient")
  
  # Add dateok.bin column
  predSympWide$dateok.bin <- case_when(predSympWide$dateok <= ymd("2009-01-22") ~ 0, 
                                       predSympWide$dateok > ymd("2009-01-22") ~ 1)

  predSympWide$dateok.bin <- factor(predSympWide$dateok.bin, levels = c(0, 1)) # 0 = before/on 2009-01-22; 1 = after 2009/01/22
}

# Add additional columns for bAEplaqueCombine
if(str_detect(pattern, "AEplaqueCombine")) {
  cat("Bulk RNA-seq data of AEplaqueCombine detected...")
  
  # Add batch ID
  batchesPlaque <- readRDS("data/metadata/AEplaque_batch_patientID.rds")
  
  predSympWide <- predSympWide %>%
    left_join(batchesPlaque, by = "patient")
  
  predSympLong <- predSympLong %>%
    left_join(batchesPlaque, by = "patient")
}

rm(predModLong, predModWide)
```

```{r}
# ---- Impute missing values ----
# sum(is.na(predSympWide$MMP2))

contCol <- c("BMI", "GFR_MDRD", "FABP4", "VEGFA_plasma") # MODIFY HERE
binCol <- c("SmokerCurrent", "PAOD", "CAD_history", "Med.statin.derived", "Calc.bin", "Macrophages.bin", "SMC.bin", "thrombus.bin", "IPH.bin", "collagen.bin") # MODIFY HERE

# Prepare columns for imputation
contColToImpute <- predSympWide %>%
  select(patient, all_of(contCol)) %>%
  mutate(across(all_of(contCol), as.numeric))

groupColToImpute <- predSympWide %>%
  select(patient, all_of(binCol))

toImpute <- left_join(contColToImpute, groupColToImpute, by = "patient")

# Define imputation methods: 'pmm' for continuous, 'logreg' for binary
meth <- make.method(toImpute)
meth[contCol] <- "pmm"
meth[binCol] <- "logreg"

colImputedList <- mice(toImpute, method = meth, seed = 2468)
colImputed <- complete(colImputedList, 1)

# Replace column with one containing imputed values
predSympWideImputed <- predSympWide

table(predSympWide$patient == predSympWideImputed$patient)

for(col in colnames(toImpute)[-1]) {
  predSympWideImputed[[col]] <- colImputed[[col]]
}

# sum(is.na(predSympWideImputed$Calc.bin))
```

# 5.3 Clinical analysis

## 5.3.1 Simple statistical tests

```{r}
# Test for normality
resShapiro <- predSympLong %>%
  group_by(cellType) %>%
  summarise(pvalue = shapiro_test(proportion)$p.value) %>%
  add_significance("pvalue")

for(celltype in unique(predSympLong$cellType)) {
  subset <- predSympLong %>%
    filter(cellType == celltype)

  qqp <- ggpubr::ggqqplot(subset$proportion) +
    labs(title = paste("Q-Q plot for", celltype))

  print(qqp)
}

# Test for homogeneity of variance
resLevene <- predSympLong %>%
  levene_test(proportion ~ cellType)
```

```{r}
charCols <- c("Symptoms.2g", "Symptoms.4g", "Gender", "Macrophages.bin", "SMC.bin", "IPH.bin", "smc_macrophages_ratio", "plaquephenotype")

# Order cell types based on proportion to make sure that the significance bars will match the right boxplot in Section 5.4.1.
orderedCelltypes <- predSympLong %>%
  group_by(cellType) %>%
  summarise(proportion = mean(proportion)) %>%
  arrange(proportion) %>%
  pull(cellType)

predSympLong$cellType <- factor(predSympLong$cellType, levels = orderedCelltypes)

predSympLongFilter <- sympFilter(predSympLong, charCols)

resStats <- list()
resDunnList <- list()

for(symp in names(predSympLongFilter)) {
  df <- predSympLongFilter[[symp]]
  
  formula <- as.formula(sprintf("proportion ~ %s", symp))

  if(n_distinct(df[[symp]]) > 2) {
    res <- df %>%
      group_by(cellType) %>%
      kruskal_test(formula = formula) %>%
      add_significance("p")
    
    if(any(res$p <= 0.05)) {
      resDunn <- df %>%
        group_by(cellType) %>%
        dunn_test(formula = formula, p.adjust.method = "none") %>%
        add_significance("p") %>%
        add_xy_position(x = "cellType")
    
        resDunnList[[symp]] <- resDunn
      }
  } else {
    res <- df %>%
      group_by(cellType) %>%
      wilcox_test(formula = formula) %>% 
      add_significance("p") %>%
      add_xy_position(x = "cellType")
  }
  resStats[[symp]] <- res
}
```

## 5.3.2 General Linear Model

```{r}
# Variables to be used in the fitting of the GLM
responseVar <- "plaquephenotype.bin" # MODIFY HERE
covariates <- c("Age", "Gender", "BMI") # MODIFY HERE; for interaction terms, add together in one string ("var1 * var2")
glmFamily <- "binomial" # MODIFY HERE; dependent on responseVar

useProportion <- TRUE # MODIFY HERE; use estimated cellular proportions (TRUE) or bulk RNA-seq gene counts (FALSE) for GLM

# Assess whether GLM will be done with cell type proportions or gene counts and prepare the data accordingly
if(useProportion) {
  cellTypes <- colnames(pred)[2:ncol(pred)]
  glmType <- "cellTypes"
} else {
  geneMarkers <- colnames(bulkFilterMarkers)[2:ncol(bulkFilterMarkers)]
  glmType <- "geneMarkers"
}

# Generate formulas to be used
formulasGLM <- sapply(get(glmType), function(x) {
  if(length(covariates) > 0) {
    as.formula(sprintf("%s ~ %s + %s", responseVar, x, paste(covariates, collapse = " + ")))
  } else {
    as.formula(sprintf("%s ~ %s", responseVar, x))
  }
})

# Fit GLM
resGLM <- lapply(formulasGLM, function(f) {
  glModel <- tryCatch(glm(f, family = glmFamily, data = predSympWideImputedGene), 
                       error = function(e) {
                         message("GLM fitting failed for formula: ", deparse(f))
                       })
})

# lapply(resGLM, function(r) summary(r))

# Extract results from all fitted models
tidyResGLM <- lapply(seq_along(resGLM), function(i) {
  tblRes <- tidy(resGLM[[i]], conf.int = TRUE) %>%
    mutate(sig = stars.pval(p.value),
           {{glmType}} := factor(names(resGLM)[i]))
})

tblAllResGLM <- do.call(rbind, tidyResGLM)

tblSubsetResGLM <- tblAllResGLM %>%
  filter(!!sym(glmType) == term) %>%
  mutate(p.adj = p.adjust(p.value, method = "fdr"),
         sig.adj = stars.pval(p.adj)) %>%
  select(!!sym(glmType), p.value, p.adj, sig, sig.adj, estimate, conf.low, conf.high, everything(), -term)

if(glmFamily == "binomial") {
  tblAllResGLM <- mutate(tblAllResGLM, OR = exp(estimate))
}

# Extract model performance metrics from all fitted models
performanceResGLM <- lapply(resGLM, function(r) {
  tblGlance <- glance(r)
})

tblAllPerfGLM <- do.call(rbind, performanceResGLM)
```

```{r}
# ---- Validate GLM results ----
if (glmFamily == "gaussian") {
  if (length(covariates) > 0) {
    # Check for multicollinearity among covariates
    dfCor <- predSympWide %>%
      select(covariates) %>%
      na.omit()
    
    ggcorrplot::ggcorrplot(cor(dfCor), hc.order = TRUE, type = "lower", lab = TRUE)
  }

  # Test for normality of residuals
  residualShapiro <- lapply(resGLM, function(r) { 
    res <- shapiro.test(residuals(r))
    res$p.value # p.value <= 0.05 may indicate bad model fit
  })

} else if (glmFamily == "binomial") {
  if (length(covariates) > 0) {
    # Check for multicollinearity among covariates
    vifRes <- lapply(resGLM, function(r) { 
      vif(r) # VIF > 5 is problematic # NOTE TO SELF: gvif() use for glm??
    })
  }
  
  # Test for normality of residuals
  performance_hosmer(resGLM[[1]], n_bins = 10)
}

# Residual plots
lapply(names(resGLM), function(name) {
  par(mfrow = c(2, 2))
  plot(resGLM[[name]], main = name)
})
```

## 5.3.3 Cox Regression Model

```{r}
# Variables to be used in the Cox regression
timeVar <- "EP_major_time" # MODIFY HERE
statusVar <- "EP_major" # MODIFY HERE
covariates <- c("batch", "BMI", "Age", "Gender") # MODIFY HERE
useProportion <- TRUE # MODIFY HERE; use estimated cellular proportions (TRUE) or bulk RNA-seq gene counts (FALSE) for Cox regression

# Assess whether Cox regression will be done with cell type proportions or gene counts and prepare the data accordingly
if(useProportion) {
  cellTypes <- colnames(pred)[2:ncol(pred)]
  coxType <- "cellTypes"
} else {
  geneMarkers <- colnames(bulkFilterMarkers)[2:ncol(bulkFilterMarkers)]
  coxType <- "geneMarkers"
}

# Generate formulas to be used
formulasCox <- sapply(get(coxType), function(x) {
  if(length(covariates) > 0) {
    as.formula(sprintf("Surv(%s, %s) ~ %s + %s", timeVar, statusVar, x, paste(covariates, collapse = " + ")))
  } else {
    as.formula(sprintf("Surv(%s, %s) ~ %s", timeVar, statusVar, x))
  }
})

# Fit Cox proportional hazards model
resCox <- lapply(formulasCox, function(f) {
  coxModel <- tryCatch(coxph(f, data = predSympWideImputedGene), 
                       error = function(e) {
                         message("Cox regression failed for formula: ", deparse(f))
                       })
})

# lapply(resCox, function(r) summary(r))

# Extract results from all fitted models
tidyResCox <- lapply(seq_along(resCox), function(i) {
  tblRes <- tidy(resCox[[i]], conf.int = TRUE, exponentiate = TRUE) %>%
    mutate(sig = stars.pval(p.value),
           {{coxType}} := factor(names(resCox)[i]))
})

tblAllResCox <- do.call(rbind, tidyResCox)

tblSubsetResCox <- tblAllResCox %>%
  filter(!!sym(coxType) == term) %>%
  mutate(p.adj = p.adjust(p.value, method = "fdr"),
         sig.adj = stars.pval(p.adj)) %>%
  rename(HR = estimate) %>%
  select(!!sym(coxType), p.value, p.adj, sig, sig.adj, HR, conf.low, conf.high, everything(), -term)

performanceResCox <- lapply(resCox, function(x) {
  tblGlance <- glance(x)
})

tblAllPerfCox <- do.call(rbind, performanceResCox) %>%
  rownames_to_column({{coxType}}) %>%
  select({{coxType}}, AIC, logLik, everything())
```

```{r}
# ---- Validate Cox regression  results ----
# Check for multicollinearity among covariates
dfToCor <- predSympWideImputedGene %>%
  mutate_at(covariates, as.numeric) %>%
  select(all_of(covariates))
  
ggcorrplot::ggcorrplot(cor(dfToCor), hc.order = TRUE, type = "lower", lab = TRUE)

# Test Cox proportional hazards assumption
testCox <- lapply(resCox, cox.zph)

print(testCox)
```

# 5.4 Graphs

## 5.4.1 Boxplot of predicted cellular proportion, divided on patient characteristics

```{r}
labelSymp <- list(Symptoms.2g = c("mild", "severe"),
                  Symptoms.4g = c("asymptomatic", "ocular", "tia", "stroke"),
                  Gender = c("female", "male"),
                  Macrophages.bin = c("no/minor", "moderate/heavy"),
                  SMC.bin = c("no/minor", "moderate/heavy"),
                  IPH.bin = c("no", "yes"),
                  smc_macrophages_ratio = c("smc dominant", "equal", "macrophages dominant"),
                  plaquephenotype = c("fibrous", "fibroatheromatous", "atheromatous")) # MODIFY HERE

listFilterPredSympLong <- list()

for(symp in names(labelSymp)) {
  filterPredSympLong <- predSympLong %>%
    filter(!is.na(!!sym(symp)) & !!sym(symp) != -888 & !!sym(symp) != -999) %>%
    select(patient, cellType, proportion, !!sym(symp))
  
  listFilterPredSympLong[[symp]] <- filterPredSympLong
}
```

```{r}
# ---- Without significance bars ----
plotList <- list()

for(symp in names(labelSymp)) {
  nLabels <- as.character(length(labelSymp[[symp]]))
  
  pSymp <- ggplot(listFilterPredSympLong[[symp]], aes(x = reorder(cellType, proportion), y = proportion)) +
  geom_boxplot(aes(fill = as.factor(!!sym(symp))), position = position_dodge(width = 0.8), width = 0.5, linewidth = 0.3, outlier.size = 1, outlier.shape = 1) +
  labs(title = pattern,
       x = "",
       y = "Estimated cell type proportion",
       fill = symp) +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, by = 0.1)) +
  scale_fill_manual(values = blueColorPal[[nLabels]], labels = labelSymp[[symp]]) + # MODIFY HERE; custom color palette options in functions_figures_R.R
  theme_classic(base_family = "Helvetica") +
  boxplotTheme()
  
  plotList[[symp]] <- pSymp
}

print(plotList)

path <- paste0("output/plots/boxplot_cell_type_proportion_symptoms/", pattern)

createDir(path)

for(name in names(plotList)) {
  tiff(paste0(path, "/", name, "_", pattern, ".tiff"))
  
  print(plotList[[name]])
  
  dev.off()
}
```

```{r}
# ---- With significance bars ----
signPlotList <- list()

for(symp in names(resStats)) {
  if(any(resStats[[symp]]$p.signif != "ns")) {
    if(symp %in% names(resDunnList)) {
      df <- resDunnList[[symp]]
    } else {
      df <- resStats[[symp]]
    }
    
    pAdd <- plotList[[symp]]
    
    pSign <- pAdd +
      stat_pvalue_manual(df, label = "p.signif", hide.ns = TRUE, tip.length = 0.5)
    
    signPlotList[[symp]] <- pSign
  }
}

print(signPlotList)

pathSign <- paste0("output/plots/boxplot_cell_type_proportion_symptoms_significant/", pattern)

if(length(signPlotList) != 0) {
  createDir(pathSign)
  for(name in names(signPlotList)) {
    tiff(paste0(pathSign, "/", name, "_", pattern, ".tiff"))
    
    print(signPlotList[[name]])
    
    dev.off()
    }
  } else {
    cat("List is empty")
}
```

## 5.4.2 Forest plot

### 5.4.2.1 GLM

```{r}
plotAllResGLM <- tblSubsetResGLM %>%
  mutate(sig.bin = ifelse(sig != " " & sig != ".", "yes", "no"),
         sig.adj.bin = ifelse(sig.adj != " " & sig.adj != ".", "yes", "no"))

plotAllResGLM$sig.bin <- factor(plotAllResGLM$sig.bin, levels = c("no", "yes"))
plotAllResGLM$sig.adj.bin <- factor(plotAllResGLM$sig.adj.bin, levels = c("no", "yes"))

plotPAdj <- FALSE # MODIFY HERE; use the adjusted p.values (TRUE) or non-adjusted (FALSE)

colPBin <- ifelse(plotPAdj, "sig.adj.bin", "sig.bin")

glmPlotTitle <- ifelse(plotPAdj, paste(gsub("B_cell", "cellType", paste0(deparse(formulasGLM[[1]]))), "(adj)"), gsub("B_cell", "cellType", paste0(deparse(formulasGLM[[1]]))))

pGLM <- ggplot(plotAllResGLM, aes(x = !!rlang::sym(glmType), y = estimate)) +
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high, color = !!rlang::sym(colPBin)), 
                width = NA, linewidth = 0.4, show.legend = TRUE) +
  geom_point(aes(color = !!rlang::sym(colPBin), shape = !!rlang::sym(colPBin)), 
             size = 2, stroke = NA, show.legend = TRUE) +
  geom_hline(yintercept = 0, linetype = "dotted", linewidth = 0.2) +
  coord_flip() +
  labs(x = "",
       y = expression(beta ~ "coefficient"),
       color = expression(bold(significant))) +
  scale_color_manual(values = c("black", "#4A97CA"), drop = FALSE) +
  scale_fill_manual(values = c("black", "#4A97CA"), drop = FALSE) +
  scale_shape_manual(values = c(16, 15), drop = FALSE) +
  guides(color = guide_legend(override.aes = list(
                                  shape = c(16, 15),
                                  linewidth = 0.6,
                                  color = c("black", "#4A97CA"))),
         fill = "none",
         shape = "none") +
  theme_classic(base_family = "Helvetica",
                base_line_size = 0.2) +
  theme(plot.title = element_text(size = 5, face = "bold"),
        axis.ticks.length = unit(0.07, "cm"),
        axis.title = element_text(size = 6),
        axis.text = element_text(size = 6),
        axis.text.x = element_text(margin = margin(t = 2)),
        axis.text.y = element_text(margin = margin(r = 2)),
        legend.position = "top",
        legend.margin = margin(b = -8),
        legend.title = element_text(size = 6),
        legend.text = element_text(size = 6),
        legend.key.size = unit(0.4, "cm"),
        aspect.ratio = 1)

basePath <- ifelse(plotPAdj, "output/plots/glm_forest_plot/adjusted/", "output/plots/glm_forest_plot/")

# tiff(paste0(basePath, "scaden_", pattern, "_glm_", responseVar, "_", paste0(covariates, collapse = "_"),".svg", collapse = ""), 
#      width = 7, height = 10, unit = "cm", res = 300)
# 
# print(pGLM)
# 
# dev.off()
```

### 5.4.2.2 Cox regression model

```{r}
plotAllResCox <- tblSubsetResCox %>%
  mutate(HR_CI = sprintf("%.2f [%.2f - %.2f]", HR, conf.low, conf.high),
         sig.bin = ifelse(sig != " " & sig != ".", "yes", "no"),
         sig.adj.bin = ifelse(sig.adj != " " & sig.adj != ".", "yes", "no"))

plotAllResCox$sig.bin <- factor(plotAllResCox$sig.bin, levels = c("no", "yes"))
plotAllResCox$sig.adj.bin <- factor(plotAllResCox$sig.adj.bin, levels = c("no", "yes"))

plotPAdj <- FALSE # MODIFY HERE; use the adjusted p.values (TRUE) or non-adjusted (FALSE)

if(useProportion) {
  baseFormulaCox <- paste0(statusVar, " ~ cellType", ifelse(!is.null(covariates), paste(" +", covariates, collapse = "+"), ""))
} else {
  baseFormulaCox <- paste0(statusVar, " ~ gene", ifelse(!is.null(covariates), paste(" +", covariates, collapse = "+"), ""))
}

if(plotPAdj) {
  colP <- "p.adj"
  colPBin <- "sig.adj.bin"
  colPSig <- "sig.adj"
  pPValueTitle <- "p adj"
  baseFormulaCox <- paste(baseFormulaCox, "(adj)")
} else {
  colP <- "p.value"
  colPBin <- "sig.bin"
  colPSig <- "sig"
  pPValueTitle <- "p"
}

pLabelsTitle <- str_to_sentence((gsub("([A-Z])", " \\1", coxType)))
```

```{r}
# https://rpubs.com/mbounthavong/forest_plots_r
# Create the base plot for the different text columns
baseCol <- ggplot(plotAllResCox, aes(y = !!rlang::sym(coxType))) +
  theme(text = element_text(family = "Helvetica"),
        plot.title = element_text(size = 9, face = "bold", vjust = -1),
        axis.text.x = element_text(color = "white", hjust = -3),
        axis.line = element_blank(),
        axis.title.x = element_blank(),
        axis.text.y = element_blank(), 
        axis.ticks = element_blank(),
        axis.title.y = element_blank(), 
        legend.position = "none",
        panel.background = element_blank(), 
        panel.border = element_blank(), 
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), 
        plot.background = element_blank())

pLabels <- baseCol +
  geom_text(aes(x = -Inf, label = !!rlang::sym(coxType)), hjust = 0, size = 2.5, family = "Helvetica") +
  ggtitle(pLabelsTitle) +
  annotate("text", x = 0, y = Inf, label = "") +
  theme(plot.title = element_text(hjust = 0))

pForest <- ggplot(plotAllResCox, aes(x = !!rlang::sym(coxType), y = HR)) +
  geom_hline(yintercept = 1, linewidth = 0.3, color = "#878787") +
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high, color = !!rlang::sym(colPBin)), width = NA, linewidth = 0.3) +
  geom_point(aes(color = !!rlang::sym(colPBin), shape = !!rlang::sym(colPBin)), size = 1, stroke = 0.8) +
  coord_flip() +
  labs(x = "",
       y = expression(bold("Hazard Ratio")),
       color = expression(bold(significant))) +
  scale_color_manual(values = c("black", "#4A97CA")) + # Colors for the outlines
  scale_fill_manual(values = c("black", "#4A97CA")) + # Colors for the squares
  scale_shape_manual(values = c(16, 15)) +
  guides(color = "none",
         fill = "none",
         shape = "none") +
  theme_classic() +
  theme(plot.title = element_text(size = 7),
        axis.text.x = element_text(family = "Helvetica", size = 7),
        axis.title.x = element_text(family = "Helvetica", size = 7),
        axis.line.y = element_blank(),
        axis.ticks.y= element_blank(),
        axis.text.y= element_blank(),
        axis.title.y= element_blank(),
        axis.ticks.length.x = unit(0.2, "cm"),
        axis.line = element_line(color = "#878787", linewidth = 0.3),
        axis.ticks = element_line(color = "#878787", linewidth = 0.3),
        panel.background = element_blank(),
        plot.background = element_blank())

pHR <- baseCol +
  geom_text(aes(x = 0, label = HR_CI), hjust = 0.5, size = 2.5, family = "Helvetica") +
  ggtitle("HR [95% CI]") +
  theme(plot.title = element_text(hjust = 0.5))

pPValue <- baseCol +
  geom_text(aes(x = 0, label = sprintf("%.2f", !!rlang::sym(colP))), hjust = 0.5, size = 2.5, family = "Helvetica") +
  ggtitle(pPValueTitle) +
  theme(plot.title = element_text(hjust = 0.5))

pSig <- baseCol +
  geom_text(aes(x = 0, label = !!rlang::sym(colPSig)), hjust = 0.5, size = 2.5, family = "Helvetica") +
  ggtitle("Sig.") +
  theme(plot.title = element_text(hjust = 0.5))

pLabels + pForest + pHR + pPValue + pSig + plot_layout(design = c(area(t = 0, l = 0, b = 5, r = 39),
                                                                    area(t = 0, l = 36, b = 5, r = 57),
                                                                    area(t = 0, l = 57, b = 5, r = 72),
                                                                    area(t = 0, l = 69, b = 5, r = 80),
                                                                    area(t = 0, l = 80, b = 5, r = 80)))

basePath <- "output/plots/cox_forest_plot/"

addPath <- paste0(basePath, 
                  ifelse(plotPAdj, "adjusted/", "non_adjusted/"),
                  ifelse(useProportion, "cell_type_proportion/", "gene_marker_count/"))

ggsave(paste0(addPath, "scaden_", pattern, "_", paste0(covariates, collapse = "_"), ".svg", collapse = ""), 
     width = 10, height = 10, units = "cm", dpi = 300)
```
